Updates for maven

+++ javaUpdatesOne.sh
@@ -14,8 +14,10 @@ if [[ -n $latestJavaLTS ]]; then
   done
 
   f=".java-version"
-  sed -E "s/[0-9]*/$latestJavaLTS/g" "$f" > "$f""01"
-  mv "$f""01" "$f"
+  if [ -f $f ]; then
+    sed -E "s/[0-9]*/$latestJavaLTS/g" "$f" > "$f""01"
+    mv "$f""01" "$f"
+  fi
 
   for f in $(find . -name "build.gradle*"); do
       sed -E 's/java\.sourceCompatibility\s*=\s*JavaVersion\.VERSION_[0-9]*/java.sourceCompatibility = JavaVersion.VERSION_'"$latestJavaLTS"'/g' "$f" > "$f""01"
@@ -23,6 +25,13 @@ if [[ -n $latestJavaLTS ]]; then
       sed -E "s/jvmTarget\s*=\s* \"[0-9]*\"/jvmTarget = \"$latestJavaLTS\"/g" "$f" > "$f""01"
       mv "$f""01" "$f"
       sed -E "s/languageVersion = JavaLanguageVersion.of\([0-9]*\)/languageVersion = JavaLanguageVersion.of($latestJavaLTS)/g" "$f" > "$f""01"
+      mv "$f""01" "$f"
+  done
+
+  for f in $(find . -name "pom.xml"); do
+      sed -E "s/<java\.version>[0-9]*<\/java\.version>/<java.version>$latestJavaLTS<\/java.version>/g" "$f" > "$f""01"
+      mv "$f""01" "$f"
+      sed -E "s/<java\.jdk>[0-9]*<\/java\.jdk>/<java.jdk>$latestJavaLTS<\/java.jdk>/g" "$f" > "$f""01"
       mv "$f""01" "$f"
   done
 else
